Implement the filtering functionality.
There is already an existing UI element that is a drop down menu for filtering based on names.  With this change, when a particular name is selected, the page gets reloaded with the corresponding filtering applied.  The special value of `All` means no filtering.  The special value of `Surprise me` means that either Igor or Jasmine will be picked.

This implementation is rather simplistic and could be more generic when needed.

I'll send out as a pull request in case there is additional feedback.  Thanks for considering this change!